+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.2 (unreleased)
 
-- Added `CrossNet` and `DeepFM` modules
+- Added `CrossNet`, `DeepFM`, and `FactorizationMachine` modules
 
 ## 0.0.1 (2022-02-28)